# api/storage/docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Virta Stations API",
        "contact": {
            "email": "mail@alexseman.com"
        },
        "version": "0.0.1"
    },
    "paths": {
...
        "/stations": {
            "get": {
                "tags": [
                    "stations",
                    "index",
                    "get"
                ],
                "summary": "Stations Index",
                "description": "Returns a paginated index of the stations",
                "operationId": "stationsIndex",
                "parameters": [
                    {
                        "name": "page",
                        "in": "query",
                        "required": false
                    },
                    {
                        "name": "per_page",
                        "in": "query",
                        "required": false
                    },
                    {
                        "name": "all",
                        "in": "query",
                        "description": "Overrides pagination and returns all the stations for use for /stations/by-company",
                        "required": false
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ApiResource"
                                }
                            }
                        }
                    }
                }
            },
            "post": {
                "tags": [
                    "stations",
                    "store",
                    "post"
                ],
                "summary": "Stations Store/Create",
                "description": "Returns a list of stations that meet the search criteria",
                "operationId": "stationsStore",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "": {
                                        "properties": {
                                            "name": {
                                                "description": "Station name",
                                                "type": "string"
                                            },
                                            "address": {
                                                "description": "Station physical address",
                                                "type": "string"
                                            },
                                            "lat": {
                                                "description": "Latitude part of the station coordinates",
                                                "type": "number",
                                                "format": "float"
                                            },
                                            "long": {
                                                "description": "Longitude part of the station coordinates",
                                                "type": "number",
                                                "format": "float"
                                            },
                                            "company_id": {
                                                "description": "The company ID of the station",
                                                "type": "integer"
                                            }
                                        },
                                        "type": "object"
                                    }
                                },
                                "type": "object",
                                "example": {
                                    "name": "Riverside Lake 1",
                                    "address": "3943 Winfield, Agustinaport, RI 15842-8736",
                                    "lat": 122.704904,
                                    "long": 71.593214,
                                    "company_id": 15
                                }
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Success",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean"
                                        },
                                        "message": {
                                            "type": "string"
                                        },
                                        "data": {
                                            "type": "object",
                                            "oneOf": [
                                                {
                                                    "type": "StationResource"
                                                }
                                            ]
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validation Error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean"
                                        },
                                        "errors": {
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean"
                                        },
                                        "errors": {
                                            "type": "array",
                                            "items": {
                                                "type": "string"
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/stations/search": {
            "get": {
                "tags": [
                    "stations",
                    "search",
                    "get"
                ],
                "summary": "Stations Search",
                "description": "Returns a list of stations that meet the search criteria",
                "operationId": "stationsSearch",
                "parameters": [
                    {
                        "name": "lat",
                        "in": "query",
                        "description": "Latitude for the starting point criteria of the stations search",
                        "required": true
                    },
                    {
                        "name": "long",
                        "in": "query",
                        "description": "Longitude for the starting point criteria of the stations search",
                        "required": true
                    },
                    {
                        "name": "radius",
                        "in": "query",
                        "description": "Indicates the KM radius of the search that is to be performed around the starting point",
                        "required": true
                    },
                    {
                        "name": "company_id",
                        "in": "query",
                        "description": "If present the search will be performed only for the stations of the given company and its eventual subsidiaries",
                        "required": false
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ApiResource"
                                }
                            }
                        }
                    },
                    "204": {
                        "description": "Successful operation but no results found meeting the criteria",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ApiResource"
                                }
                            }
                        }
                    }
                }
            }
        },
...
    },
    "components": {
        "schemas": {
            "ApiResource": {
                "title": "ApiResource",
                "properties": {
                    "success": {
                        "type": "boolean"
                    },
                    "message": {
                        "type": "string"
                    },
                    "errors": {
                        "type": "array",
                        "items": {
                            "type": "string"
                        }
                    },
                    "data": {
                        "type": "array",
                        "items": {
                            "anyOf": [
                                {
                                    "type": "CompanyResource"
                                },
                                {
                                    "type": "StationResource"
                                }
                            ]
                        }
                    }
                },
                "type": "object"
            },
...
        }
    }
}